+++ Source/Geometry/Plane/Plane.h
@@ -8,7 +8,7 @@ public:
 	Plane(const Point& _p, const Vector& _n, const RGBColor& _c);
 	~Plane();
 
-	bool HitTest(const Ray& _r, float& _t, const float& _tMin) override;
+	bool HitTest(const Ray& _r, double& _t, const float& _tMin) override;
 
 protected:
 	Point mCenter;
